[Refact] #140 - 카페 전체 조회 (토큰인증 O)
- CafeService의 getListCafes를 CafeAuthService로 이동
- 서비스 모듈에 CafeAuthService 추가
