+++ core/network/src/main/java/org/cazait/network/datasource/CafeListRemoteData.kt
@@ -1,6 +1,7 @@
 package org.cazait.network.datasource
 
 import org.cazait.network.NetworkConnectivity
+import org.cazait.network.api.auth.CafeAuthService
 import org.cazait.network.api.auth.FavoriteService
 import org.cazait.network.api.unauth.CafeService
 import org.cazait.network.di.Authenticated
@@ -16,6 +17,7 @@ import javax.inject.Inject
 
 class CafeListRemoteData @Inject constructor(
     private val cafeService: CafeService,
+    @Authenticated private val cafeAuthService: CafeAuthService,
     @Authenticated private val favoriteCafeService: FavoriteService,
     private val networkConnectivity: NetworkConnectivity,
 ) : CafeListRemoteDataSource {
@@ -24,10 +26,10 @@ class CafeListRemoteData @Inject constructor(
         query: ListCafesReq
     ): DataResponse<ListCafesRes> {
         return when (val response = processCall {
-            cafeService.getListCafes(
+            cafeAuthService.getListCafes(
                 userId = userId,
-                latitude = query.latitude,
-                longitude = query.longitude,
+                query.longitude,
+                query.latitude,
                 sort = query.sort,
                 limit = query.limit
             )
