+++ core/network/src/main/java/org/cazait/network/di/ServiceModule.kt
@@ -4,6 +4,7 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import org.cazait.network.api.auth.CafeAuthService
 import org.cazait.network.api.auth.FavoriteService
 import org.cazait.network.api.unauth.AuthService
 import org.cazait.network.api.auth.ReviewService
@@ -57,13 +58,13 @@ object ServiceModule {
     ): FavoriteService {
         return retrofit.create(FavoriteService::class.java)
     }
-//
-//    @Provides
-//    @Singleton
-//    @Authenticated
-//    fun providesAuthServiceAuth(
-//        @Authenticated retrofit: Retrofit
-//    ): AuthService {
-//        return retrofit.create(AuthService::class.java)
-//    }
+
+    @Provides
+    @Singleton
+    @Authenticated
+    fun providesCafeAuthServiceAuth(
+        @Authenticated retrofit: Retrofit
+    ): CafeAuthService {
+        return retrofit.create(CafeAuthService::class.java)
+    }
 }
